+++ react/components/FieldInput.tsx
@@ -1,49 +1,62 @@
 import type { FC, SyntheticEvent } from 'react'
 import React from 'react'
-import { Input } from 'vtex.styleguide'
-
-interface DataLocaleTypes {
-  [key: string]: string
-}
+import { Input, Checkbox } from 'vtex.styleguide'
+import { FormattedMessage } from 'react-intl'
 
 interface FieldInputProps {
   binding: Binding
-  dataLocales: DataLocaleTypes
   handleChange: (e: SyntheticEvent) => void
-  key: number
-  showBindings: { [key: string]: boolean }
-  showEditValue: BindingTranslation[]
   showValue: boolean
+  translationInfo: BindingTranslation
+  handleHideLabel: ({
+    bindingId,
+    status,
+  }: {
+    bindingId: string
+    status: boolean
+  }) => void
 }
 
 const FieldInput: FC<FieldInputProps> = (props: FieldInputProps) => {
   const {
     binding,
-    dataLocales,
     handleChange,
-    key,
-    showBindings,
     showValue,
+    handleHideLabel,
+    translationInfo,
   } = props
 
-  const label =
-    dataLocales[binding.id] && showValue ? dataLocales[binding.id] : ''
-
   return (
-    <div key={key} className="flex items-center justify-center w-100">
+    <div className="flex items-center justify-center w-100">
       <div className="pa4 w-40">
         <label>
           {binding.canonicalBaseAddress} ({binding.defaultLocale})
         </label>
       </div>
-      <div className="pa4 w-50">
+      <div className="pa4 w-50 flex items-center">
         <Input
-          disabled={!showBindings[binding.id]}
-          required={showBindings[binding.id]}
+          disabled={!showValue || translationInfo.hide}
+          required={showValue || !translationInfo.hide}
           name={binding.id}
           onChange={(e: SyntheticEvent) => handleChange(e)}
-          value={label}
+          value={translationInfo.label}
         />
+        <div className="ml3 flex items-center">
+          <Checkbox
+            disabled={!showValue}
+            id={`hide-label-${binding.id}`}
+            checked={!!translationInfo.hide}
+            onChange={() =>
+              handleHideLabel({
+                status: !translationInfo.hide,
+                bindingId: binding.id,
+              })
+            }
+          />
+          <label htmlFor={`hide-label-${binding.id}`} className="ml3">
+            <FormattedMessage id="hide-label" />
+          </label>
+        </div>
       </div>
     </div>
   )