+++ react/components/FormDialog.tsx
@@ -1,12 +1,16 @@
 import type { SyntheticEvent, FormEvent, FC } from 'react'
-import React, { useState, useEffect, useCallback } from 'react'
+import React, { useState, useEffect } from 'react'
 import { FormattedMessage } from 'react-intl'
 import { Modal, Button } from 'vtex.styleguide'
 import { useMutation } from 'react-apollo'
 
 import saveBindingInfo from '../graphql/saveBindingInfo.gql'
-import FieldInput from './FieldInput'
-import { setShowValues } from '../utils'
+import {
+  removeBindingAdmin,
+  setShowValues,
+  createBindingsToLabel,
+} from '../utils'
+import AdminBindingLabelsList from './AdminBindingLabelsList'
 
 interface FormDialogProps {
   open: boolean
@@ -15,10 +19,6 @@ interface FormDialogProps {
   chosenBinding: Binding
   bindingInfoQueryData: BindingsSaved[]
   setFetchedData: (binding: BindingsSaved[]) => void
-}
-
-interface DataLocaleTypes {
-  [key: string]: string
 }
 
 interface DataMutation {
@@ -35,141 +35,78 @@ const FormDialog: FC<FormDialogProps> = (props: FormDialogProps) => {
     setFetchedData,
   } = props
 
-  const [dataLocales, setDataLocales] = useState<DataLocaleTypes>({})
-  const [saveTranslatedInfo] = useMutation<BindingsSaved>(saveBindingInfo)
-  const [translatedLocales, setTranslatedLocales] = useState<
-    BindingTranslation[]
-  >([])
+  const [saveTranslatedInfo] = useMutation<BindingsSaved, DataMutation>(
+    saveBindingInfo
+  )
+
+  const [translationsMap, setTranslationMap] = useState<
+    Record<string, BindingTranslation>
+  >({})
 
   const showBindings = setShowValues(bindingInfoQueryData)
 
-  const getTranslatedLabels = useCallback(
-    (arr: BindingsSaved[]) => {
-      const filtered = arr.filter(
-        (binding: { bindingId: string }) =>
-          binding.bindingId === chosenBinding.id
-      )
-
-      return filtered[0]?.translatedLocales
-    },
-    [chosenBinding.id]
-  )
-
   useEffect(() => {
-    setFetchedData(bindingInfoQueryData)
-    const translatedLabels = getTranslatedLabels(bindingInfoQueryData ?? [])
+    const bindingsToLabel = createBindingsToLabel(
+      bindings,
+      bindingInfoQueryData.find(
+        (binding) => binding.bindingId === chosenBinding.id
+      )?.translatedLocales ?? []
+    )
 
-    setTranslatedLocales(translatedLabels ?? [])
-
-    const [chosenBindingData] =
-      bindingInfoQueryData.filter(
-        (bind) => bind.bindingId === chosenBinding.id
-      ) ?? []
-
-    const initialLabels = {} as DataLocaleTypes
-
-    if (chosenBindingData?.translatedLocales?.length) {
-      chosenBindingData?.translatedLocales?.forEach((b) => {
-        initialLabels[b.id] = b.label ? b.label : ''
-      })
-    }
-
-    setDataLocales(initialLabels)
-  }, [
-    chosenBinding.id,
-    getTranslatedLabels,
-    bindingInfoQueryData,
-    setFetchedData,
-  ])
+    setTranslationMap(bindingsToLabel)
+  }, [bindingInfoQueryData, bindings, chosenBinding.id])
 
   const handleChange = (event: SyntheticEvent) => {
-    const { name, value } = event.target as HTMLButtonElement
+    const { name, value } = event.target as HTMLInputElement
 
-    setDataLocales({ ...dataLocales, [name]: value })
+    setTranslationMap({
+      ...translationsMap,
+      [name]: {
+        ...translationsMap[name],
+        label: value,
+      },
+    })
   }
 
   const onSubmit = (e: FormEvent) => {
     e.preventDefault()
-    const payload = {} as BindingsSaved
-    const dataContainer = {} as DataMutation
 
-    payload.bindingId = chosenBinding.id
-    const translatedInfoArray = [] as BindingTranslation[]
+    const transformedData = bindingInfoQueryData.map<BindingsSaved>(
+      (binding) => {
+        if (binding.bindingId === chosenBinding.id) {
+          return {
+            ...binding,
+            translatedLocales: Object.values(translationsMap),
+          }
+        }
 
-    for (const [key, value] of Object.entries(dataLocales)) {
-      const [
-        {
-          defaultLocale,
-          canonicalBaseAddress,
-          extraContext: {
-            portal: { salesChannel },
-          },
-        },
-      ] = bindings.filter(({ id }) => id === key)
+        return binding
+      }
+    )
 
-      translatedInfoArray.push({
-        label: value,
-        id: key,
-        defaultLocale,
-        canonicalBaseAddress,
-        salesChannel: salesChannel.toString(),
-      })
-    }
+    saveTranslatedInfo({ variables: { data: transformedData } })
 
-    payload.translatedLocales = translatedInfoArray
-    payload.show = !!showBindings[chosenBinding.id]
+    setFetchedData(transformedData)
 
-    if (bindingInfoQueryData.length) {
-      const filteredFromChosenId = bindingInfoQueryData.filter(
-        (item: { bindingId: string }) => item.bindingId !== chosenBinding.id
-      )
-
-      filteredFromChosenId.push(payload)
-
-      dataContainer.data = filteredFromChosenId
-    } else {
-      const newArray = [] as BindingsSaved[]
-
-      newArray.push(payload)
-      dataContainer.data = newArray
-    }
-
-    saveTranslatedInfo({ variables: dataContainer })
-
-    setFetchedData(dataContainer.data)
-    const translatedLabels = getTranslatedLabels(dataContainer.data)
-
-    setTranslatedLocales(translatedLabels ?? [])
     handleOnClose()
   }
 
-  const showFields = () => {
-    const fields = bindings
-      ?.filter((binding: Binding) => {
-        return binding.canonicalBaseAddress.split('/')[1] !== 'admin'
-      })
-      .map((binding: Binding, i: number) => {
-        const [showBinding] =
-          bindingInfoQueryData.filter(
-            (bind) => bind.bindingId === binding.id
-          ) ?? []
+  const bindingsToBeLabeled = removeBindingAdmin(bindings)
 
-        const showValue: boolean = showBinding?.show
-
-        return (
-          <FieldInput
-            binding={binding}
-            dataLocales={dataLocales}
-            handleChange={handleChange}
-            key={i}
-            showBindings={showBindings}
-            showEditValue={translatedLocales ?? []}
-            showValue={showValue}
-          />
-        )
-      })
-
-    return fields
+  const handleHideLabel = ({
+    bindingId,
+    status,
+  }: {
+    bindingId: string
+    status: boolean
+  }) => {
+    setTranslationMap({
+      ...translationsMap,
+      [bindingId]: {
+        ...translationsMap[bindingId],
+        hide: status,
+      },
+    })
   }
 
   return (
@@ -184,7 +121,13 @@ const FormDialog: FC<FormDialogProps> = (props: FormDialogProps) => {
           <FormattedMessage id="admin-modal" />
         </div>
         <div className="pt6 flex w-100 flex-column justify-center items-center">
-          {showFields()}
+          <AdminBindingLabelsList
+            bindings={bindingsToBeLabeled}
+            activeBindings={showBindings}
+            handleChange={handleChange}
+            translationsMap={translationsMap}
+            handleHideLabel={handleHideLabel}
+          />
           <div className="flex pt6">
             <div className="pr4">
               <Button variation="tertiary" onClick={() => handleOnClose()}>